PlaneDAA: fix feature-disabled / no-double / param CI failures
Integration gaps in the PlaneDAA branch exposed by the full CI matrix
(test_build_options, no-double boards, per-vehicle param XML), now that the
branch is on current upstream:

- ADSB_ENABLE "missing in XML" (Heli/QuadPlane/PlaneTests1a): the @Param block
  above AP_GROUPINFO("ENABLE",16,...) was mislabelled @Param: OPTIONS, so XML
  documented ADSB_OPTIONS twice and never ADSB_ENABLE. Fix the name (kept the
  real ADSB_OPTIONS untouched). Param stays available on all vehicles.

- SPRacingH7 DO_NOT_USE_DOUBLE_MATHS: AP_Avoidance.cpp used double fabs() in the
  scripting helpers; switch to fabsf() (lines 860, 914).

- CubeOrange "'logger' is not a member of 'AP'": the FENC dev-log block in
  AC_Fence::check() was guarded by `#if 1`; guard with `#if HAL_LOGGING_ENABLED`.

- Pixhawk6X "'AP_OAScripting' does not name a type" / -Werror=undef: the
  ArduPlane aoscripting member and AP_OAScripting's enable macro weren't gated
  consistently. Make AP_OA_SCRIPTING_ENABLED ALWAYS defined (fold in
  AP_SCRIPTING_ENABLED, so `#if` is valid under -Werror=undef), include
  AP_OAScripting.h unconditionally in Plane.h (header self-guards), and gate the
  member on AP_OA_SCRIPTING_ENABLED. The .cpp keeps an explicit APM_BUILD_TYPE so
  waf still compiles it per-vehicle.

Verified: SPRacingH7 copter (fabs); sitl plane default/--disable-scripting/
--disable-AP_AVOIDANCE all build clean (no does-not-name-a-type, no -Wundef, no
@Invalid_use_of_APM_BUILD_TYPE); param_parse emits ADSB_ENABLE.

+++ libraries/AP_Avoidance/AP_Avoidance.cpp
@@ -857,7 +857,7 @@ float AP_Avoidance::distance_to_obstacle(const Vector3f &start_NED_m, const Vect
             distance_m = distance_m - get_obstacle_radius_m(obstacle.emitter_type);
 
             // height difference is the minimum difference in the height between the line and the obstacle
-            float height_difference_m = MIN(fabs(start_NED_m.z - obstacle_NED_m.z), fabs(end_NED_m.z - obstacle_NED_m.z));
+            float height_difference_m = MIN(fabsf(start_NED_m.z - obstacle_NED_m.z), fabsf(end_NED_m.z - obstacle_NED_m.z));
 
             if (distance_m < distance_new_m && height_difference_m < get_obstacle_height_m(obstacle.emitter_type)) {
                 // we are within the horizontal distance - next check the vertical distance
@@ -911,7 +911,7 @@ float AP_Avoidance::distance_to_aircraft(const Vector3f &vehicle_NED_m, const fl
             float distance_msq = (vehicle_NE_m - obstacle_NE_m).length_squared();
 
             // height difference is the difference in the height between the vehicle and the obstacle
-            float height_difference_m = fabs(vehicle_NED_m.z - obstacle_NED_m.z);
+            float height_difference_m = fabsf(vehicle_NED_m.z - obstacle_NED_m.z);
 
             // this finds the nearest aircraft iff it is within the height limits for 
             if (distance_msq < distance_new_msq && height_difference_m < get_obstacle_height_m(obstacle.emitter_type)) {
